Refactoring user modal
